Conflicts:
	projects/OG-Financial/src/main/java/com/opengamma/financial/analytics/model/credit/isda/calibration/ISDACDSHazardRateCurveFunction.java
	projects/OG-Financial/src/main/java/com/opengamma/financial/analytics/model/credit/isda/calibration/ISDACDXAsSingleNameHazardRateCurveFunction.java

+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/financial/instrument/future/BondFuturesTransactionDefinition.java
@@ -26,7 +26,7 @@ public class BondFuturesTransactionDefinition extends FuturesTransactionDefiniti
    * @param tradeDate Transaction date.
    * @param tradePrice Transaction price.
    */
-  public BondFuturesTransactionDefinition(final BondFuturesSecurityDefinition underlyingFuture, final int quantity, final ZonedDateTime tradeDate, final double tradePrice) {
+  public BondFuturesTransactionDefinition(final BondFuturesSecurityDefinition underlyingFuture, final long quantity, final ZonedDateTime tradeDate, final double tradePrice) {
     super(underlyingFuture, quantity, tradeDate, tradePrice);
   }
